Move two more scripts out and put plotting modules in sub-package
